// src/types/user.ts
export interface Users {
    id: string;
    firstName: string;
    lastName: string;
    password: string;
    address: string;
    avatar: string;
    city: string;
    country: string;
    email: string;
    createdAt: string;
    postalCode: string;
    role: string;
    phone: string;
    name: string;
  }